+++ azure_devops_rust_api/examples/git_pr_create.rs
@@ -1,0 +1,65 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+// git_pr_create.rs
+// Create Pull Request example.
+use anyhow::Result;
+use azure_devops_rust_api::git;
+use azure_devops_rust_api::Credential;
+use git::models::GitPullRequestCreateOptions;
+use std::env;
+use std::sync::Arc;
+
+#[tokio::main]
+async fn main() -> Result<()> {
+    // Initialize logging
+    env_logger::init();
+
+    // Get authentication credential either from a PAT ("ADO_TOKEN") or via the az cli.
+    let credential = match env::var("ADO_TOKEN") {
+        Ok(token) => {
+            println!("INFO:Authenticate using PAT provided via $ADO_TOKEN");
+            Credential::from_pat(token)
+        }
+        Err(_) => {
+            println!("INFO:Authenticate using Azure CLI");
+            Credential::from_token_credential(Arc::new(azure_identity::AzureCliCredential {}))
+        }
+    };
+
+    const USAGE: &str =
+        "Usage: git_pr_create <repository-name> <src_branch> <target_branch> <title> <description>";
+
+    // Get ADO server configuration via environment variables
+    let organization = env::var("ADO_ORGANIZATION").expect("Must define ADO_ORGANIZATION");
+    let project = env::var("ADO_PROJECT").expect("Must define ADO_PROJECT");
+
+    let repo_name = env::args().nth(1).expect(USAGE);
+    let src_branch: String = env::args().nth(2).expect(USAGE);
+    let target_branch: String = env::args().nth(3).expect(USAGE);
+    let title: String = env::args().nth(4).expect(USAGE);
+    let description: String = env::args().nth(5).expect(USAGE);
+
+    // Create a git client
+    let git_client = git::ClientBuilder::new(credential).build();
+
+    // Create GitPullRequestCreateOptions with all the mandatory parameters
+    let mut pr_create_options = GitPullRequestCreateOptions::new(
+        // Need to specify full git refs path
+        format!("refs/heads/{src_branch}"),
+        format!("refs/heads/{target_branch}"),
+        title,
+    );
+    // Set any additional optional parameters
+    pr_create_options.description = Some(description);
+
+    // Define the new PR
+    let pr = git_client
+        .pull_requests_client()
+        .create(organization, repo_name, project, pr_create_options)
+        .into_future()
+        .await?;
+    println!("Created PR:\n{:#?}", pr);
+
+    Ok(())
+}